完成语音发送功能

+++ app/src/main/java/com/example/jinphy/simplechat/services/common_service/aidl/service/DownloadFileBinder.java
@@ -4,10 +4,14 @@ import android.content.Context;
 
 import com.example.jinphy.simplechat.models.file_task.FileTask;
 import com.example.jinphy.simplechat.models.file_task.FileTaskRepository;
+import com.example.jinphy.simplechat.models.message.Message;
+import com.example.jinphy.simplechat.models.message.MessageRepository;
+import com.example.jinphy.simplechat.modules.system_msg.new_member.NewMemberAdapter;
 import com.example.jinphy.simplechat.services.common_service.aidl.models.FileDownloader;
 import com.example.jinphy.simplechat.services.common_service.aidl.models.OnError;
 
 import static com.example.jinphy.simplechat.broadcasts.AppBroadcastReceiver.TAG_DOWNLOAD_FILE;
+import static com.example.jinphy.simplechat.broadcasts.AppBroadcastReceiver.TAG_DOWNLOAD_VOICE;
 import static com.example.jinphy.simplechat.broadcasts.AppBroadcastReceiver.TAG_UPLOAD_FILE;
 import static com.example.jinphy.simplechat.broadcasts.AppBroadcastReceiver.send;
 
@@ -19,13 +23,25 @@ import static com.example.jinphy.simplechat.broadcasts.AppBroadcastReceiver.send
 public class DownloadFileBinder extends IDownloadFileBinder.Stub {
 
 
-    private final Context context;
+    private Context context;
 
     private FileTaskRepository fileTaskRepository;
+    private MessageRepository messageRepository;
 
-    public DownloadFileBinder(Context context) {
-        this.context = context;
+    private static class InstanceHolder{
+        static final DownloadFileBinder DEFUALT = new DownloadFileBinder();
+    }
+
+    public static DownloadFileBinder getInstance(Context context) {
+        InstanceHolder.DEFUALT.context = context;
+        return InstanceHolder.DEFUALT;
+    }
+
+
+
+    public DownloadFileBinder() {
         fileTaskRepository = FileTaskRepository.getInstance();
+        messageRepository = MessageRepository.getInstance();
     }
 
 
@@ -46,6 +62,22 @@ public class DownloadFileBinder extends IDownloadFileBinder.Stub {
                 })
                 .doOnFinish(id -> {
                     send(context, TAG_DOWNLOAD_FILE, "onFinish:" + id);
+                    fileTaskRepository.remove(fileTaskId);
+                })
+                .execute();
+    }
+
+    @Override
+    public void downloadVoice(long fileTaskId, long msgId) {
+        FileTask task = fileTaskRepository.get(fileTaskId);
+        FileDownloader uploader = FileDownloader.with(task);
+        if (uploader == null) {
+            send(context, TAG_DOWNLOAD_VOICE, Message.AUDIO_STATUS_NO+":"+msgId);
+            return;
+        }
+        uploader.doOnError(id -> send(context, TAG_DOWNLOAD_VOICE, Message.AUDIO_STATUS_NO+":"+msgId))
+                .doOnFinish(id -> {
+                    send(context, TAG_DOWNLOAD_VOICE, Message.AUDIO_STATUS_NEW + ":" + msgId);
                     fileTaskRepository.remove(fileTaskId);
                 })
                 .execute();
